bug fix to fadeinfadeout animation and put in new file
